Add files via upload
In this assignment, you are asked to develop a Java program which
computes the convex hull or convex envelope of N points using Graham’s scan
algorithm. The convex hull of N points is the smallest convex polygon that
contains all the points of it. It finds the minimum area convex region that
contains every point.